+++ gcgc/tokenizer/kmer_tokenzier.py
@@ -5,9 +5,10 @@
 from typing import List, Optional
 import itertools as it
 
-from pydantic import Field
+from pydantic import Field, validator
 
 from gcgc.tokenizer.base import SequenceTokenizer, SequenceTokenizerSettings
+from gcgc import alphabets
 
 
 class KmerTokenizerSettings(SequenceTokenizerSettings):
@@ -19,9 +20,14 @@ class KmerTokenizerSettings(SequenceTokenizerSettings):
 
     max_length: Optional[int] = Field(None, env="GCGC_MAX_LENGTH")
 
+    @validator("alphabet")
+    def resolve_alphabet(cls, alphabet):  # pylint: disable=no-self-use, no-self-argument
+        """Resolve the alphabet if it's a named alphabet."""
+        return alphabets.resolve_alphabet(alphabet)
+
     @property
     def special_tokens(self) -> List[str]:
-        """Returns the list of special tokens that are not None."""
+        """Return the list of special tokens that are not None."""
         return [
             x
             for x in [
@@ -58,36 +64,34 @@ def _create_kmer_vocab_from_token(
 class KmerTokenizer(SequenceTokenizer):
     """A sequence tokenizer."""
 
-    def __init__(self, tokenizer_settings: KmerTokenizerSettings):
+    def __init__(self, settings: KmerTokenizerSettings):
         """Init the SequenceTokenizer class.
 
         Args:
-            tokenizer_settings: The settings for the tokenizer.
+            settings: The settings for the tokenizer.
             vocabulary: The vocabulary for the tokenizer.
 
         """
-        self.tokenizer_settings = tokenizer_settings
+        self.settings = settings
         self.vocab = _create_kmer_vocab_from_token(
-            self.tokenizer_settings.alphabet,
-            self.tokenizer_settings.kmer_length,
-            self.tokenizer_settings.special_tokens,
+            self.settings.alphabet, self.settings.kmer_length, self.settings.special_tokens,
         )
 
     def _kmer_n(self, seq: str) -> List[str]:
         seq_len = len(seq)
-        iterations = seq_len - self.tokenizer_settings.kmer_length + 1
+        iterations = seq_len - self.settings.kmer_length + 1
         kmers = []
 
-        for i in range(0, iterations, self.tokenizer_settings.kmer_stride):
-            kmer = seq[i : i + self.tokenizer_settings.kmer_length]
+        for i in range(0, iterations, self.settings.kmer_stride):
+            kmer = seq[i : i + self.settings.kmer_length]
             kmers.append(kmer)
 
         return kmers
 
-    def encode_as_ids(self, seq: str) -> List[int]:
+    def encode(self, seq: str) -> List[int]:
         """Encode the underlying sequence into a list of tokens."""
         return [
-            self.vocab.get(s, self.vocab.get(self.tokenizer_settings.unk_token))
+            self.vocab.get(s, self.vocab.get(self.settings.unk_token))
             for s in self.encode_as_tokens(seq)
         ]
 
@@ -103,24 +107,24 @@ class KmerTokenizer(SequenceTokenizer):
         """
         seq_len = len(seq)
 
-        if seq_len < self.tokenizer_settings.kmer_length:
+        if seq_len < self.settings.kmer_length:
             raise ValueError(
                 f"seq length {seq_len} cannot be less than the kmer "
-                "size {self.tokenizer_settings.kmer_length}"
+                "size {self.settings.kmer_length}"
             )
 
-        if self.tokenizer_settings.kmer_length == 1:
+        if self.settings.kmer_length == 1:
             kmer_list = list(seq)
         else:
             kmer_list = self._kmer_n(seq)
 
-        if self.tokenizer_settings.bos_token:
-            kmer_list = [self.tokenizer_settings.bos_token] + kmer_list
+        if self.settings.bos_token:
+            kmer_list = [self.settings.bos_token] + kmer_list
 
-        if self.tokenizer_settings.eos_token:
-            kmer_list = kmer_list + [self.tokenizer_settings.eos_token]
+        if self.settings.eos_token:
+            kmer_list = kmer_list + [self.settings.eos_token]
 
-        if self.tokenizer_settings.max_length:
-            return kmer_list[: self.tokenizer_settings.max_length]
-        else:
-            return kmer_list
+        if self.settings.max_length:
+            return kmer_list[: self.settings.max_length]
+
+        return kmer_list